Moved ID details from node API to commands
