--- api/id.go
@@ -1,103 +1,0 @@
-package api
-
-import (
-	"encoding/base64"
-	"encoding/json"
-
-	ma "gx/ipfs/QmNTCey11oxhb1AxDnQBRHtdhap6Ctud872NjAYPYYXPuc/go-multiaddr"
-	"gx/ipfs/QmTu65MVbemtUxJEWgsTtzv9Zv9P8rvmqNA4eG9TrTRGYc/go-libp2p-peer"
-)
-
-// IDDetails is a collection of information about a node.
-type IDDetails struct {
-	Addresses       []ma.Multiaddr
-	ID              peer.ID
-	AgentVersion    string
-	ProtocolVersion string
-	PublicKey       []byte // raw bytes
-}
-
-// ID is the interface that defines methods to fetch identifying information
-// about the underlying node.
-type ID interface {
-	// Details, returns detailed information about the underlying node.
-	Details() (*IDDetails, error)
-}
-
-// MarshalJSON implements json.Marshaler
-func (idd IDDetails) MarshalJSON() ([]byte, error) {
-	addressStrings := make([]string, len(idd.Addresses))
-	for i, addr := range idd.Addresses {
-		addressStrings[i] = addr.String()
-	}
-
-	v := map[string]interface{}{
-		"Addresses": addressStrings,
-	}
-
-	if idd.ID != "" {
-		v["ID"] = idd.ID.Pretty()
-	}
-	if idd.AgentVersion != "" {
-		v["AgentVersion"] = idd.AgentVersion
-	}
-	if idd.ProtocolVersion != "" {
-		v["ProtocolVersion"] = idd.ProtocolVersion
-	}
-	if idd.PublicKey != nil {
-		// Base64-encode the public key explicitly.
-		// This is what the built-in JSON encoder does to []byte too.
-		v["PublicKey"] = base64.StdEncoding.EncodeToString(idd.PublicKey)
-	}
-	return json.Marshal(v)
-}
-
-// UnmarshalJSON implements Unmarshaler
-func (idd *IDDetails) UnmarshalJSON(data []byte) error {
-	var v map[string]*json.RawMessage
-	var err error
-	if err = json.Unmarshal(data, &v); err != nil {
-		return err
-	}
-
-	var addresses []string
-	if err := decode(v, "Addresses", &addresses); err != nil {
-		return err
-	}
-	idd.Addresses = make([]ma.Multiaddr, len(addresses))
-	for i, addr := range addresses {
-		a, err := ma.NewMultiaddr(addr)
-		if err != nil {
-			return err
-		}
-		idd.Addresses[i] = a
-	}
-
-	var id string
-	if err := decode(v, "ID", &id); err != nil {
-		return err
-	}
-	if idd.ID, err = peer.IDB58Decode(id); err != nil {
-		return err
-	}
-
-	if err := decode(v, "AgentVersion", &idd.AgentVersion); err != nil {
-		return err
-	}
-	if err := decode(v, "ProtocolVersion", &idd.ProtocolVersion); err != nil {
-		return err
-	}
-	if err := decode(v, "PublicKey", &idd.PublicKey); err != nil {
-		return err
-	}
-	return nil
-}
-
-func decode(idd map[string]*json.RawMessage, key string, dest interface{}) error {
-	if raw := idd[key]; raw != nil {
-		if err := json.Unmarshal(*raw, &dest); err != nil {
-			return err
-		}
-	}
-	return nil
-}
