+++ tools/fast/action_id.go
@@ -2,12 +2,13 @@ package fast
 
 import (
 	"context"
-	"github.com/filecoin-project/go-filecoin/api"
+
+	"github.com/filecoin-project/go-filecoin/commands"
 )
 
 // ID runs the `id` command against the filecoin process
-func (f *Filecoin) ID(ctx context.Context, options ...ActionOption) (*api.IDDetails, error) {
-	var out api.IDDetails
+func (f *Filecoin) ID(ctx context.Context, options ...ActionOption) (*commands.IDDetails, error) {
+	var out commands.IDDetails
 	args := []string{"go-filecoin", "id"}
 
 	for _, option := range options {
